Player career stats

NBA player career statistics for specified players and inputs

get_players_career_stats(players = NULL, player_ids = NULL,
  modes = c("PerGame", "Totals"), assign_to_environment = TRUE,
  add_mode_names = TRUE, return_message = TRUE)

Arguments

players
NULL or vector of NBA players

player_ids
vector of NBA Player IDs

modes
vector of items that can include

Totals

PerGame

Per36

assign_to_environment
if TRUE assign to each table environment with a name starting with data

add_mode_names
if TRUE adds the type of mode to the table data

return_message
if TRUE returns a message

Value

a data_frame

See also

Other player: get_bref_players_bios, get_players_awards, get_players_tables_data, get_years_draft_combines

Examples

# NOT RUN {
get_players_career_stats(players = c("Joe Harris", "Myles Turner", "Spencer Dinwiddie"),
modes = c("Totals", "PerGame"))
# }